Drop unsupported Python versions
We're not supporting Python 2.7, as `oemof` decided to not do so and
we're heavily using that.
Python 3.7 is a problem, because `pyproj`, a dependency of
`omoef.tabular`, has to be slightly modified and built from source for
this. Since the goal is to eventually support Python 3.7, I left its
build environments in `tox.ini` but decided to ignore errors on this for
now. This means that builds our build is passing for all supported
versions, but we can still continually monitor the build behaviour in
order to see when it eventually works.
PyPy is Python 2.7 so that gets dropped anyway. Whether PyPy3 works or
not, has to be checked via a continuous integration service, as I don't
have that interpreter at hand.