Implement hierarchical scenarios
This is done by implementing the `Scenario` class, which has the
following features:

  * It encapsulates loading scenario from a path via the `from_path`
    class method and
  * it looks for a `parents` key in the loaded scenario, which enables
    copying other scenarios into this one.

The `from_path` method does everything that is necessary to load a
scenario from a file. This currently entails:

  * using a special function from `oemof.tabular` instead of just using
    `toml.load` and
  * registering the loaded scenario as known under `fuchur.scenarios`
    under its name and its path.

A scenario's `parents` can be specified as a list of strings, which will
be looked up in `fuchur`'s known scenarios. If they string is not found,
it is tried to be loaded via `Scenario.from_path`.